add parameter location enum
